Docs: explain what 'flasher-3.5' means
Explicitly document the need to install a flasher tool and point to the
download sites. 'flasher-3.5' isn't a well known command, and with the
demise of the official distribution sites, it's become hard to find.

Also describe how to do the same thing with 0xFFFF, the free software
replacement for maemo_flasher.